refactor: mejora estructura carpetas (front)
