+++ frontend/src/components/imc/calculadora/components/Categorias.tsx
