+++ frontend/src/components/imc/calculadora/components/Input.tsx
@@ -13,7 +13,7 @@ import {
   FormMessage,
 } from "@/components/ui/form"
 import { Input } from "@/components/ui/input"
-import { ImcFormValues, imcSchema } from "./utils/imcSchema"
+import { ImcFormValues, imcSchema } from "../../utils/imcSchema"
 
 interface ImcInputProps {
   onSubmit: (values: ImcFormValues) => void
