+++ frontend/src/components/imc/calculadora/components/Resultado.tsx
