+++ frontend/src/components/imc/calculadora/hooks/useCalculator.ts
